moves period generators
