+++ lib/period.dart
@@ -17,11 +17,10 @@
 library period;
 
 import 'src/enums/enums.dart' show PeriodGenerator;
-import 'src/period_generator.dart';
+import 'src/period_generators/period_generators.dart';
 import 'src/periods/periods.dart';
 
 export 'src/enums/enums.dart' show Month, PeriodGenerator, Week, Weekday;
 export 'src/everies/everies.dart';
-export 'src/extensions/extensions.dart' show EndOfDay;
-export 'src/period_generator.dart';
+export 'src/period_generators/period_generators.dart';
 export 'src/periods/periods.dart';
