+++ lib/src/enums/weekday.dart
@@ -1,7 +1,7 @@
 import 'package:time/time.dart';
 import '../date_validators/date_validators.dart';
 import '../everies/everies.dart';
-import '../period_generator.dart';
+import '../period_generators/period_generators.dart';
 import '../periods/periods.dart';
 
 /// Weekday constants that are returned by [DateTime.weekday] method.
