+++ lib/src/extensions/extensions.dart
@@ -2,7 +2,6 @@ export 'add_days.dart';
 export 'clamp_in_month.dart';
 export 'date_validator_list_ext.dart';
 export 'day_in_year.dart';
-export 'end_of_day.dart';
 export 'every_date_validator_list_ext.dart';
 export 'previous_next.dart';
 export 'week_calc.dart';
