+++ lib/src/periods/week_period.dart
@@ -1,7 +1,6 @@
 import 'package:time/time.dart';
 import '../enums/enums.dart';
-import '../extensions/extensions.dart';
-import '../period_generator.dart';
+import '../period_generators/period_generators.dart';
 import 'day_period.dart';
 import 'day_period_bundle.dart';
 import 'period.dart';
