+++ lib/src/periods/year_period.dart
@@ -1,6 +1,6 @@
 import 'package:time/time.dart';
-import '../extensions/extensions.dart';
-import '../period_generator.dart';
+import '../period_generators/month_generator.dart';
+import '../period_generators/period_generators.dart';
 import 'month_period.dart';
 import 'month_period_bundle.dart';
 import 'period.dart';
